microservices/car_detector/requirements.txt
flask
ultralytics
kafka-python
opencv-python
numpy
pandas
